+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - arg_parser.arg_file_chk:  Replaced exception code with call to _create_file function.
 
 ### Added
+- arg_parser._parse_single:  Private function for use by arg_parser2 function.
 - arg_parser._file_create:  Private function for use by arg_file_chk function.
 - arg_parser._parse_multi:  Private function for use by arg_parser2 function.